knowflow/server/requirements.txt
Flask==3.1.0
flask_cors==5.0.1
flask_jwt_extended==4.6.0
mysql-connector-python==9.2.0
pycryptodomex==3.20.0
tabulate==0.9.0
Werkzeug==3.1.3
PyJWT==2.10.1
dotenv==0.9.9
transformers==4.51.3
elasticsearch==8.12.0
minio==7.2.4
strenum==0.4.15
peewee==3.17.1
pytz==2020.5
tiktoken==0.9.0
ragflow-sdk==0.18.0
markdown==3.6
markdown-to-json==2.1.1
markdown-it-py[linkify,plugins]==3.0.0
geopy
Pillow
PyYAML
pydantic
openpyxl
pymysql>=1.1.0
loguru
pandas
pdf2image